Completley redesigned GUI Interface
I have gone through and completley redesigned the User Interface GUI
for the touch panel. Rather than having a single screen view, we now
have a three-screen view that can now be navigated using buttons.
Currently, the Calibration window needs to be edited such that the
CAL and ZERO buttons properly COM with the NXP. Data Saving to a USB
also needs to be implemented